move RedisDriver to redis_driver.py
